removed SampleFormat, added Mixer helper class SampleFormat was an unnecessary parameter, since theres no need for 32bit audio

+++ src/main/java/com/jbatista/wmo/filter/HighPass.java
@@ -9,7 +9,7 @@ import com.jbatista.wmo.util.MathFunctions;
 
 public class HighPass extends BiquadFilter {
 
-    public HighPass(double sampleRate) {
+    public HighPass(int sampleRate) {
         this.sampleRate = sampleRate;
         setCutoffFrequency(440);
         setResonance(1);
